+++ src/ChessDotCpp/attacks.h
@@ -10,4 +10,5 @@ public:
 	static Bitboard GetAttackersOf(Board board, Position position, Bitboard allPieces);
 	static Bitboard GetAttackersOfSide(const Board& board, Position position, bool byWhite, Bitboard allPieces);
 	static Bitboard GetCheckers(const Board& board);
+	static bool IsPositionAttacked(const Board& board, const Position position, const bool byWhite);
 };